requirements.txt
django
django-dashing
requests
ipdb